Split out Time into separate types for Time values and DateMath times
The Time type has historically been used to represent both Elasticsearch's TimeValue and time within Date Math expressions. This is slightly misleading because

1. TimeValue accepts units from nanos -> seconds
2. Date Math expressions with time accept units from seconds -> years

In both cases, Elasticsearch accepts only whole numbers in addition to a unit of measure.

This commit introduces a new DateMathTime type to represent times within DateMath expressions, with a DateMathTimeUnit enum to represent the valid units for this type. There are implicit conversions from string, TimeSpan and double such that implicit conversion usage in users upgrading from 5.x -> 6.x will not be affected by the change. The constructors that accept string, TimeSpan and double accept fractional numbers to make the type easier to work with, and the fractional number will be converted to the largest whole number and unit that can be represented, with rounding to the nearest second for fractional seconds. An optional rounding argument to determine how to round mid points can be passed to the constructors that accept TimeSpan, double and string.

The DateMathTime type implements equality and comparison members, comparing the seconds representation. An approximation of seconds is used in the case of years and months, with 365 * seconds in a day for years, and (365 / 12)  * seconds in a day for months, allowing for comparisons between months and years.

Remove support for units above a day from the Time type.

Closes #2983